+++ src/main/java/io/github/manuelarte/spring/manuelartevalidation/validators/ExistsValidator.java
@@ -2,15 +2,16 @@ package io.github.manuelarte.spring.manuelartevalidation.validators;
 
 import io.github.manuelarte.spring.manuelartevalidation.constraints.Exists;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.StreamSupport;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
+import javax.validation.constraints.NotNull;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.core.ResolvableType;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Component;
-import org.springframework.util.Assert;
 
 @Component
 @Scope(value = ConfigurableBeanFactory.SCOPE_PROTOTYPE)
@@ -18,27 +19,37 @@ import org.springframework.util.Assert;
 public class ExistsValidator implements ConstraintValidator<Exists, Object> {
 
   @lombok.NonNull
-  private final List<CrudRepository> repositories;
-  private CrudRepository repository;
+  private final List<CrudRepository<?, ?>> repositories;
+  private CrudRepository<Object, Object> repository;
 
   @Override
-  public void initialize(final Exists exists) {
-    this.repository = repositories.stream().filter(it -> ResolvableType.forClass(it.getClass())
-        .as(CrudRepository.class).getGeneric(0).resolve().equals(exists.value()))
-        .findFirst().orElseThrow(() -> new RuntimeException("Can't find the repository for the document"));
+  public void initialize(@NotNull final Exists exists) {
+    this.repository = (CrudRepository<Object, Object>) repositories.stream().filter(it -> Objects
+        .requireNonNull(ResolvableType.forClass(it.getClass())
+            .as(CrudRepository.class).getGeneric(0).resolve()).equals(exists.value()))
+        .findFirst().orElse(null);
   }
 
   @Override
   public boolean isValid(final Object id, final ConstraintValidatorContext cxt) {
-    Assert.isTrue( id instanceof Iterable || id.getClass().equals(repositoryIdClass()),
-        String.format("%s is not the expected id type of %s", id.getClass(), repositoryIdClass()));
+    if (repository == null) {
+      cxt.buildConstraintViolationWithTemplate("Can't find the repository for the entity");
+      return false;
+    }
+    if ( !(id == null
+        || id instanceof Iterable
+        || id.getClass().equals(repositoryIdClass()))) {
+      cxt.buildConstraintViolationWithTemplate(String.format("%s is not the expected id type of %s",
+          id.getClass(), repositoryIdClass()));
+      return false;
+    }
     return id == null
-        || !(id instanceof Iterable)
+        || (!(id instanceof Iterable)
         ? repository.existsById(id)
-        : existsIterableIds((Iterable)id);
+        : existsIterableIds((Iterable<Object>)id));
   }
 
-  private boolean existsIterableIds(Iterable<?> ids) {
+  private boolean existsIterableIds(final Iterable<Object> ids) {
     final long expectedSize = StreamSupport.stream(ids.spliterator(), false).count();
     return StreamSupport.stream(
         repository.findAllById(ids).spliterator(), false).count() == expectedSize;